#560 Moved shared HttpHeadersRule code to core package
